+++ packages/web3-providers/src/twitter/index.ts
@@ -1,141 +1,9 @@
-import { escapeRegExp } from 'lodash-es'
 import urlcat from 'urlcat'
 import LRUCache from 'lru-cache'
 import type { TwitterBaseAPI } from '../types/index.js'
+import { getSettings, getTokens, getUserNFTContainer } from './apis/index.js'
 
 const UPLOAD_AVATAR_URL = 'https://upload.twitter.com/i/media/upload.json'
-
-function getScriptURL(content: string, name: string) {
-    const matchURL = new RegExp(
-        `https://abs.twimg.com/responsive-web/\(client-web|client-web-\\w+\)\{1\}/${escapeRegExp(
-            `${name}.`,
-        )}\\w+${escapeRegExp('.js')}`,
-        'm',
-    )
-    const [url] = content.match(matchURL) ?? []
-    return url
-}
-
-function getScriptContentMatched(content: string, pattern: RegExp) {
-    const [, matched] = content.match(pattern) ?? []
-    return matched
-}
-
-function getCSRFToken() {
-    const ct0 = document.cookie.split('; ').find((x) => x.includes('ct0'))
-    if (!ct0) return ''
-    const [, value] = ct0.split('=')
-    return value
-}
-
-const scriptCache = new LRUCache<string, Promise<Response> | string>({
-    max: 10,
-    ttl: 300_000,
-})
-
-async function fetchContent(url?: string) {
-    if (!url) return
-
-    const hit: Promise<Response> = scriptCache.get(url) ?? fetch(url)
-
-    if (scriptCache.get(url) !== hit) scriptCache.set(url, hit)
-
-    if (typeof hit === 'string') return hit
-
-    const response = (await hit).clone()
-    if (!response.ok) {
-        scriptCache.delete(url)
-        return
-    }
-    const content = await response.text()
-    scriptCache.set(url, content)
-    return content
-}
-
-async function getTokens(operationName?: string) {
-    const swContent = await fetchContent('https://twitter.com/sw.js')
-    if (!swContent) throw new Error('Failed to fetch manifest script.')
-
-    const [mainContent, nftContent] = await Promise.all([
-        fetchContent(getScriptURL(swContent, 'main')),
-        fetchContent(getScriptURL(swContent, 'bundle.UserNft')),
-    ])
-
-    const bearerToken = getScriptContentMatched(mainContent ?? '', /,\w="(\w{20,}%3D\w{20,})",/)
-    const queryToken = getScriptContentMatched(nftContent ?? '', /{\s?id:\s?"([\w-]+)"/)
-    const csrfToken = getCSRFToken()
-    const queryId = operationName
-        ? getScriptContentMatched(mainContent ?? '', new RegExp(`queryId:"([^"]+)",operationName:"${operationName}"`))
-        : undefined
-
-    return {
-        bearerToken,
-        queryToken,
-        csrfToken,
-        queryId,
-    }
-}
-
-async function getUserNftContainer(
-    screenName: string,
-    queryToken: string,
-    bearerToken: string,
-    csrfToken: string,
-): Promise<{
-    data: {
-        user: {
-            result: TwitterBaseAPI.NFTContainer
-        }
-    }
-}> {
-    const response = await fetch(
-        urlcat('https://twitter.com/i/api/graphql/:queryToken/userNftContainer_Query', {
-            queryToken,
-            variables: JSON.stringify({
-                screenName,
-            }),
-            features: JSON.stringify({
-                responsive_web_twitter_blue_verified_badge_is_enabled: false,
-            }),
-        }),
-        {
-            headers: {
-                authorization: `Bearer ${bearerToken}`,
-                'x-csrf-token': csrfToken,
-                'content-type': 'application/json',
-                'x-twitter-auth-type': 'OAuth2Session',
-                'x-twitter-active-user': 'yes',
-                referer: `https://twitter.com/${screenName}/nft`,
-            },
-        },
-    )
-    return response.json()
-}
-
-async function getSettings(bearerToken: string, csrfToken: string): Promise<TwitterBaseAPI.Settings> {
-    const response = await fetch(
-        urlcat('https://twitter.com/i/api/1.1/account/settings.json', {
-            include_mention_filter: false,
-            include_nsfw_user_flag: false,
-            include_nsfw_admin_flag: false,
-            include_ranked_timeline: false,
-            include_alt_text_compose: false,
-            include_country_code: false,
-            include_ext_dm_nsfw_media_filter: false,
-        }),
-        {
-            headers: {
-                authorization: `Bearer ${bearerToken}`,
-                'x-csrf-token': csrfToken,
-                'content-type': 'application/json',
-                'x-twitter-auth-type': 'OAuth2Session',
-                'x-twitter-active-user': 'yes',
-                referer: 'https://twitter.com/home',
-            },
-        },
-    )
-    return response.json()
-}
 
 const cache = new LRUCache<string, any>({
     max: 40,
@@ -159,7 +27,7 @@ export class TwitterAPI implements TwitterBaseAPI.Provider {
     }
     async getUserNftContainer(screenName: string) {
         const { bearerToken, queryToken, csrfToken } = await getTokens()
-        const result = await getUserNftContainer(screenName, queryToken, bearerToken, csrfToken)
+        const result = await getUserNFTContainer(screenName, queryToken, bearerToken, csrfToken)
         if (!result?.data?.user?.result?.has_nft_avatar) throw new Error('No NFT avatar.')
 
         return {
